Reoganizes package layout and adds utils module.
