+++ gzip_classifier/types.py
@@ -1,0 +1,3 @@
+Model = [(str, bytes, int, str)]
+IndexRow = [((int, int), (int, int))]
+Index = [IndexRow]
